requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# PDF processing
PyPDF2>=3.0.1
PyMuPDF>=1.23.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Environment and utilities
python-dotenv>=1.0.0

# Optional: For better performance
# faiss-gpu>=1.7.4  # Uncomment if you have CUDA support
